fix int tests
