+++ docker-compose.yml
@@ -23,6 +23,7 @@ services:
       - REDIS_HOST=redis
       - AUTHORISED_ROLES=ROLE_PRISONS_REPORTING_USER
       - HMPPS_AUTH_URL=https://sign-in-dev.hmpps.service.justice.gov.uk/auth
+      - HMPPS_MANAGE_USERS_URL=https://manage-users-api-dev.hmpps.service.justice.gov.uk
       - REPORTING_API_URL=https://digital-prison-reporting-mi-dev.hmpps.service.justice.gov.uk
       - TOKEN_VERIFICATION_API_URL=https://token-verification-api-dev.prison.service.justice.gov.uk
       - TOKEN_VERIFICATION_ENABLED=true
