+++ feature.env
@@ -1,5 +1,6 @@
 PORT=3007
 HMPPS_AUTH_URL=http://127.0.0.1:9091/auth
+HMPPS_MANAGE_USERS_URL=http://127.0.0.1:9091
 TOKEN_VERIFICATION_API_URL=http://127.0.0.1:9091/verification
 TOKEN_VERIFICATION_ENABLED=true
 NODE_ENV=development
