+++ server/data/hmppsManageUsersClient.ts
@@ -13,11 +13,11 @@ export interface UserRole {
 
 export default class HmppsManageUsersClient {
   private static restClient(token: string): RestClient {
-    return new RestClient('HMPPS Auth Client', config.apis.manageUsers, token)
+    return new RestClient('HMPPS Manage User Client', config.apis.manageUsers, token)
   }
 
   getUser(token: string): Promise<User> {
-    logger.info(`Getting user details: calling HMPPS Auth`)
+    logger.info(`Getting user details: calling HMPPS Manage User Client`)
     return HmppsManageUsersClient.restClient(token).get({ path: '/users/me' }) as Promise<User>
   }
